Refactor `<TableHeadCell />`
