+++ src/components/table-head/components/cell/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useStyles } from './use-styles'
+export { useColumnDrop } from './use-column-drop'
