+++ src/components/table-head/components/cell/hooks/use-column-drop.ts
@@ -4,6 +4,21 @@
 
 import { useDrop } from 'react-dnd'
 
+export const useColumnDrop = opts => {
+    const [{ isOver, canDrop }, drop] = useDrop({
+        accept: 'HEADER',
+        // drop: drop,
+        hover: (item, mon) =>
+            handleHover(Object.assign({}, opts, { item, mon })),
+        collect: mon => ({
+            isOver: !!mon.isOver(),
+            canDrop: !!mon.canDrop()
+        })
+    })
+
+    return [drop]
+}
+
 const getColModel = (headCellRefs, columnOrder, columns) => {
     let colModel = []
     let leftMostCell = headCellRefs[0] ? headCellRefs[0] : null // left most cell is the select cell, if it exists
@@ -202,21 +217,3 @@ const handleHover = opts => {
         }
     }
 }
-
-const useColumnDrop = opts => {
-    const [{ isOver, canDrop }, drop] = useDrop({
-        accept: 'HEADER',
-        // drop: drop,
-        hover: (item, mon) =>
-            handleHover(Object.assign({}, opts, { item, mon })),
-        collect: mon => ({
-            isOver: !!mon.isOver(),
-            canDrop: !!mon.canDrop()
-        })
-    })
-
-    return [drop]
-}
-
-export { getColModel, reorderColumns, handleHover }
-export default useColumnDrop
