+++ src/components/table-head/components/cell/index.ts
@@ -1,0 +1,1 @@
+export { TableHeadCell } from './cell'
