+++ src/components/table-head/components/index.ts
@@ -1,0 +1,1 @@
+export { TableHeadCell } from './cell'
